requirements.txt
Django==3.0.8
Pillow==7.2.0
django-bootstrap4==2.2.0
django-allauth==0.42.0
django-mptt==0.11.0
